requirements.txt
boto==2.39.0
coverage==4.0.3
dj-database-url==0.4.1
dj-inmemorystorage==1.4.1
Django==1.8.12
django-crispy-forms==1.6.1
django-storages==1.4.1
flake8==3.3.0
mccabe==0.6.1
model-mommy==1.2.6
Pillow==3.2.0
pycodestyle==2.3.1
pyflakes==1.5.0
python-coveralls==2.9.1
python-decouple==3.0
PyYAML==3.12
requests==2.9.1
simplejson==3.8.2
six==1.10.0
Unipath==1.1
